app/requirements.txt
psycopg2
streamlit
python-dotenv